requirements.txt
transformers 
sentencepiece 
accelerate 
bitsandbytes
openai
pillow
opencv-python
pyyaml
diffusers
ultralytics
